Move `ILFunction` trait into `operand_iter.rs` to make it private
